split generators to appenv
